feat(mn): mnDiagram2_CreateStatRow 81.54 -> 81.81
R3-b: drop the (u8) cast on GetStatValue's return at the mnDiagram_80242B38
call. The prototype is mnDiagram_80242B38(int idx, int arg1) — the cast
minted a clrlwi r3,r3,24 the target lacks (target passes the int return
straight through). Semantically a no-op: the value is guarded < 0x19 by the
preceding check. Same m2c-era cast pattern exists in UpdateHeader:217
(noted for its own round).

Mechanism check: clrlwi census 10 -> 9 (exactly the return-mask died),
instr 427 -> 426 (target 421), save boundary/frame unchanged (stmw r22,0x50).
Protected sweep CLEAN: all 55 mndiagram* 100s hold, every partial exact.

+++ src/melee/mn/mndiagram2.c
@@ -712,8 +712,8 @@ void mnDiagram2_CreateStatRow(HSD_GObj* gobj, u8 is_name_mode, u8 stat_type,
                                                       entity_idx) < 0x19)
                     {
                         HSD_JObj* jobj = mnDiagram_80242B38(
-                            (u8) mnDiagram2_GetStatValue(
-                                is_name_mode, stat_type, entity_idx),
+                            mnDiagram2_GetStatValue(is_name_mode, stat_type,
+                                                    entity_idx),
                             0);
                         HSD_JObjSetTranslateX(jobj, -2.0f);
                         HSD_JObjSetTranslateY(jobj, (f30 * (f32) row_idx) +
